[TNT-194] refactor: 코드 정리
